Added application and user object class, created the applicationDB class, and added usability in each DB class
